# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TACHIKOMA-OS Backend API",
    "description": "Central orchestrator API for the TACHIKOMA-OS AI ecosystem. Provides LLM gateway, GraphRAG memory, chat, agent tools, and data layer for microservices.",
    "contact": {
      "name": "TACHIKOMA-OS Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/graph/export": {
      "get": {
        "tags": [
          "Graph"
        ],
        "summary": "=============================================================================",
        "description": "Export full graph (for visualization)\n=============================================================================\nGET /api/admin/graph/export\n=============================================================================",
        "operationId": "export_graph",
        "responses": {
          "200": {
            "description": "Full graph export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphExportDto"
                }
              }
            }
          },
          "500": {
            "description": "Export error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/graph/stats": {
      "get": {
        "tags": [
          "Graph"
        ],
        "summary": "=============================================================================",
        "description": "Get graph statistics\n=============================================================================\nGET /api/admin/graph/stats\n=============================================================================",
        "operationId": "get_graph_stats",
        "responses": {
          "200": {
            "description": "Graph statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphStatsDto"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "API error response",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "description": "Optional details",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "GraphEdgeDto": {
        "type": "object",
        "description": "=============================================================================\nGraph DTOs\n=============================================================================\nGraph edge response",
        "required": [
          "from_id",
          "to_id",
          "relation",
          "confidence",
          "created_at"
        ],
        "properties": {
          "confidence": {
            "type": "number",
            "format": "double",
            "description": "Confidence score"
          },
          "created_at": {
            "type": "string",
            "description": "Created timestamp"
          },
          "from_id": {
            "type": "string",
            "format": "uuid",
            "description": "Source memory ID"
          },
          "relation": {
            "type": "string",
            "description": "Relation type"
          },
          "to_id": {
            "type": "string",
            "format": "uuid",
            "description": "Target memory ID"
          }
        }
      },
      "GraphExportDto": {
        "type": "object",
        "description": "Full graph export response",
        "required": [
          "nodes",
          "edges",
          "exported_at"
        ],
        "properties": {
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphEdgeDto"
            },
            "description": "All edges"
          },
          "exported_at": {
            "type": "string",
            "description": "Export timestamp"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryDto"
            },
            "description": "All nodes"
          }
        }
      },
      "GraphStatsDto": {
        "type": "object",
        "description": "Graph statistics response",
        "required": [
          "total_nodes",
          "total_edges",
          "nodes_by_type",
          "edges_by_type",
          "avg_connections"
        ],
        "properties": {
          "avg_connections": {
            "type": "number",
            "format": "double",
            "description": "Average connections per node"
          },
          "edges_by_type": {
            "type": "object",
            "description": "Edges by type",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            }
          },
          "nodes_by_type": {
            "type": "object",
            "description": "Nodes by type",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            }
          },
          "total_edges": {
            "type": "integer",
            "description": "Total number of edges",
            "minimum": 0
          },
          "total_nodes": {
            "type": "integer",
            "description": "Total number of nodes",
            "minimum": 0
          }
        }
      },
...
      "MemoryDto": {
        "type": "object",
        "description": "=============================================================================\nMemory DTOs\n=============================================================================\nMemory node response",
        "required": [
          "id",
          "content",
          "memory_type",
          "importance_score",
          "created_at",
          "updated_at",
          "access_count"
        ],
        "properties": {
          "access_count": {
            "type": "integer",
            "format": "int64",
            "description": "Access count",
            "minimum": 0
          },
          "content": {
            "type": "string",
            "description": "Memory content"
          },
          "created_at": {
            "type": "string",
            "description": "Created timestamp"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier"
          },
          "importance_score": {
            "type": "number",
            "format": "double",
            "description": "Importance score (0.0 - 1.0)"
          },
          "memory_type": {
            "type": "string",
            "description": "Memory type"
          },
          "metadata": {
            "description": "Metadata"
          },
          "updated_at": {
            "type": "string",
            "description": "Updated timestamp"
          }
        }
      },
...
    }
  },
...
}